Create PsalmsSearcher
The Java program will ask the user for chapter and verse.

 For example:
What Book of Psalms chapter? 2
What Book of Psalms verse within the chapter? 6


The Java Program will print one of two outputs: 1) the verse text without reference, or 2) an
appropriate message if the chapter and verse does not exist